requirements.txt
Pillow>=10.0.0,<12.0.0
